test(scale): split the request-counting cases into their own file
`scale_tests.rs` reached 550 code lines and the hard limit is 500. The
split follows a seam rather than a line number: the cases that stand up
a fake libpod socket and assert how many round-trips a command makes
(#1363, #1445, #1742) move out; what stays is reconciliation logic that
touches no socket.

430 and 125 lines, both under the 300-line soft warning.

`engine_with` is duplicated rather than shared, which is what
`teardown_tests.rs` already does for the same reason: it is three lines,
and a shared helper module for it would be more machinery than the
duplication costs.
